[Jetnews] Move JetnewsSnackbar to the right folder
