+++ JetNews/app/src/main/java/com/example/jetnews/ui/components/JetnewsSnackbarHost.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.jetnews.utils
+package com.example.jetnews.ui.components
 
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.layout.wrapContentWidth
