+++ JetNews/app/src/main/java/com/example/jetnews/ui/home/HomeScreen.kt
@@ -61,8 +61,8 @@ import com.example.jetnews.data.Result
 import com.example.jetnews.data.posts.impl.BlockingFakePostsRepository
 import com.example.jetnews.model.Post
 import com.example.jetnews.ui.components.InsetAwareTopAppBar
+import com.example.jetnews.ui.components.JetnewsSnackbarHost
 import com.example.jetnews.ui.theme.JetnewsTheme
-import com.example.jetnews.utils.JetnewsSnackbarHost
 import com.example.jetnews.utils.isScrolled
 import com.example.jetnews.utils.supportWideScreen
 import com.google.accompanist.insets.LocalWindowInsets
